requirements.txt
fastai==2.7.9
Flask==2.2.2
matplotlib==3.5.3
numpy==1.23.1
Pillow==9.2.0
scikit_image==0.19.3
skimage==0.0
torch==1.12.0
torchvision==0.13.0
